clean up remainign utils
